Refactor files into a core spherical triangulation file + experiments
